refactor: 지출 또는 수입 체크박스 상태에 따른 거래내역 필터링 로직 상위 Transactions 컴포넌트로 역할 위임
