+++ frontend/src/components/Transactions/index.jsx
@@ -4,16 +4,26 @@ import { nanoid } from 'nanoid';
 
 import DailyTransaction from '../DailyTransaction';
 
+import { checkState } from '../../recoil/check/atom';
 import { transactionsInDateState } from '../../recoil/date/selector';
 
 import { TransactionsContainer } from './style';
 
 const Transactions = () => {
+    const check = useRecoilValue(checkState);
     const rawTransactions = useRecoilValue(transactionsInDateState);
+
+    let filterdTransactions = !check.income
+        ? rawTransactions.filter((transaction) => transaction.sign !== '+')
+        : rawTransactions;
+
+    filterdTransactions = !check.expenditure
+        ? filterdTransactions.filter((transaction) => transaction.sign !== '-')
+        : filterdTransactions;
 
     const dailyTransactions = new Map();
 
-    rawTransactions.forEach((rawTransaction) => {
+    filterdTransactions.forEach((rawTransaction) => {
         const currentDate = rawTransaction.date;
         if (dailyTransactions.has(currentDate)) {
             dailyTransactions.get(currentDate).push(rawTransaction);
